refact: relocate interval code
