+++ tools/canonicalize-proto-file/Main.hs
@@ -12,7 +12,6 @@
 
 module Main where
 
-import Control.Monad (guard)
 import Control.Monad.Except (runExceptT)
 
 import Data.List (sort, sortOn)
@@ -51,6 +50,7 @@ import Proto3.Suite.DotProto.AST
   , RPCMethod (..)
   )
 import Proto3.Suite.DotProto.Generate (readDotProtoWithContext)
+import Proto3.Suite.DotProto.Internal (normalizeIntervals)
 import Proto3.Suite.DotProto.Rendering (defRenderingOptions, toProtoFile)
 import Proto3.Wire.Types (FieldNumber (..))
 
@@ -66,53 +66,6 @@ data Args w = Args
 instance ParseRecord (Args Wrapped)
 
 deriving instance Show (Args Unwrapped)
-
--- | Check if the given values is some given bounds (inclusive).
-between :: (Ord a, Enum a) => a -> (a, a) -> Bool
-between x i = fst i <= x && x <= snd i
-
--- | @('isOverlappingIntervals' a b :: 'Bool')@ is a relation between two 
--- interval-like types @a@ and @b@ that is 'True' when the two intervals are 
--- overlapping or touching over a subinterval of @a@ and @b@.
-isOverlappingIntervals :: (Ord a, Enum a) => (a, a) -> (a, a) -> Bool
-isOverlappingIntervals i1@(x, y) i2@(u, v) = x `between` i2 || y `between` i2 || u `between` i1 || v `between` i1
-
--- | @('joinIntervals' a b :: 'Maybe' (a, a))@ will join the two given intervals 
--- @a@ and @b@ into a larger interval if 
---
--- @
--- ('isOverlappingIntervals' a b '==' 'True')
--- @
---
--- Otherwise, returns 'Nothing'.
-joinIntervals :: (Ord a, Enum a) => (a, a) -> (a, a) -> Maybe (a, a)
-joinIntervals a b = do 
-  guard (isOverlappingIntervals a b)
-  pure (min (fst a) (fst b), max (snd a) (snd b))
-
--- | Normalizes a list of intervals. 
---
--- 1. Filters out any "invalid" intervals from the resulting list, i.e. any 
---    intervals @(x, y) :: (a, a)@ with @x > y@. 
---
--- 2. Additionally, the resulting list will have all overlapping intervals 
---    merged into a larger interval that covers each overlapping interval.
-normalizeIntervals :: (Ord a, Enum a) => [(a, a)] -> [(a, a)]
-normalizeIntervals = mergeIntervals . filter \(x, y) -> x <= y
-
--- | Returns a sorted list that contains all intervals from the minimal set of 
--- intervals to represent the given list of intervals. 
---
--- "Merges" overlapping intervals in a list of intervals. Think disjunctive 
--- normal form.
-mergeIntervals :: (Ord a, Enum a) => [(a, a)] -> [(a, a)]
-mergeIntervals = foldr step [] . sort
-  where 
-    step :: (Ord a, Enum a) => (a, a) -> [(a, a)] -> [(a, a)]
-    step x [] = [x]
-    step x (y : ys) = case joinIntervals x y of 
-      Nothing -> x : y : ys 
-      Just xy -> xy : ys
 
 main :: IO ()
 main = do
